requirements.txt
garth>=0.4.0
requests>=2.31.0
urllib3>=2.0.0
psycopg2-binary>=2.9.0
schedule>=1.2.0
pyyaml>=6.0.0